+++ tests/testthat/test-cross_freqs.R
@@ -1,41 +1,9 @@
-# setup -------------------------------------------------------------------
-
-### Packages
-library(testthat)
-library(dplyr)
-library(y2clerk)
-library(forcats)
-library(stringr)
-
-### Data
-mtcars
-names(mtcars)
-GROUP_VARS <-
-  mtcars %>%
-  select(
-    am,
-    vs
-  ) %>%
-  names()
-
-GROUP_VARS1 <- 'am'
-GROUP_VARS2 <-
-  mtcars %>%
-  select(
-    am,
-    vs
-  ) %>%
-  names()
-
-
-
-# Tests -------------------------------------------------------------------
-### Context
-context("Tests on cross_freqs")
 
 ### Column names
 test_that("Column names", {
-  frequencies <- mtcars %>%
+  GROUP_VARS1 <- 'am'
+
+    frequencies <- mtcars %>%
     cross_freqs(
       group_vars = GROUP_VARS1,
       gear
@@ -59,13 +27,14 @@ test_that("Column names", {
 
 ### Multiple freqs vars
 test_that("Multiple freqs vars", {
+  GROUP_VARS1 <- 'am'
+
   frequencies <- mtcars %>%
     cross_freqs(
       group_vars = GROUP_VARS1,
       gear,
       vs
     )
-  frequencies
   label_levels <- unique(frequencies$label)
   variable_levels <- unique(frequencies$variable)
   expect_equal(label_levels, c("3", "4", "5", "0", "1"))
@@ -76,12 +45,13 @@ test_that("Multiple freqs vars", {
 
 ### Multiple grouping vars
 test_that("Multiple grouping vars", {
+  GROUP_VARS2 <- mtcars %>% dplyr::select(am, vs) %>% names()
+
   frequencies <- mtcars %>%
     cross_freqs(
       group_vars = GROUP_VARS2,
       gear
     )
-  frequencies
   grouping_vars <- unique(frequencies$group_var_name)
   grouping_levels <- unique(frequencies$group_var) %>% as.character()
   expect_equal(grouping_vars, c("am", "vs"))
@@ -92,13 +62,14 @@ test_that("Multiple grouping vars", {
 
 ### wide = TRUE
 test_that("wide = TRUE", {
+  GROUP_VARS2 <- mtcars %>% dplyr::select(am, vs) %>% names()
+
   frequencies <- mtcars %>%
     cross_freqs(
       group_vars = GROUP_VARS2,
       gear,
       wide = TRUE
     )
-  frequencies
   grouping_vars <- unique(frequencies$group_var_name)
   table1 <- frequencies[[2]][[1]]
   table2 <- frequencies[[2]][[2]]
@@ -113,15 +84,16 @@ test_that("wide = TRUE", {
 
 ### Mixed class group vars
 test_that("Mixed class group vars", {
+  GROUP_VARS2 <- mtcars %>% dplyr::select(am, vs) %>% names()
+
   # Long
   frequencies <- mtcars %>%
-    mutate(am = as_factor(am)) %>%
+    dplyr::mutate(am = forcats::as_factor(am)) %>%
     cross_freqs(
       group_vars = GROUP_VARS2,
       gear,
       carb
     )
-  frequencies
   grouping_vars <- unique(frequencies$group_var_name)
   grouping_levels <- unique(frequencies$group_var) %>% as.character()
   expect_equal(grouping_vars, c("am", "vs"))
@@ -129,7 +101,7 @@ test_that("Mixed class group vars", {
 
   # Wide
   frequencies <- mtcars %>%
-    mutate(am = as_factor(am)) %>%
+    dplyr::mutate(am = forcats::as_factor(am)) %>%
     cross_freqs(
       group_vars = GROUP_VARS2,
       gear,
@@ -149,7 +121,7 @@ test_that("Error messages", {
   expect_error(
     frequencies <- mtcars %>%
       cross_freqs(
-        group_vars = quos(am, vs),
+        group_vars = dplyr::quos(am, vs),
         vs
       ),
     'group_vars should be a character vector of variable names. Try formatting like c("var1", "var2") instead of c(var1, var2) or quos(var1, var2)',
@@ -170,28 +142,68 @@ test_that("Error messages", {
 
 ### exclude_groups
 test_that("exclude_groups, select() method", {
+  GROUP_VARS2 <- mtcars %>% dplyr::select(am, vs) %>% names()
+
   # Long
   frequencies <- mtcars %>%
-    select(carb, gear, am, vs) %>%
+    dplyr::select(carb, gear, am, vs) %>%
     cross_freqs(
       group_vars = GROUP_VARS2, #am & vs
       exclude_groups = TRUE
     )
-  frequencies
   grouping_vars <- unique(frequencies$variable)
   expect_equal(grouping_vars, c("carb", "gear"))
 
   # Wide
   frequencies <- mtcars %>%
-    select(carb, gear, am, vs) %>%
+    dplyr::select(carb, gear, am, vs) %>%
     cross_freqs(
       group_vars = GROUP_VARS2, #am & vs
       wide = TRUE,
       exclude_groups = TRUE
     )
-  frequencies
   grouping_vars_am <- unique(frequencies[[2]][[1]]$variable)
   grouping_vars_vs <- unique(frequencies[[2]][[2]]$variable)
   expect_equal(grouping_vars_am, c("carb", "gear"))
   expect_equal(grouping_vars_vs, c("carb", "gear"))
 })
+
+
+### include_overall
+test_that("include_overall", {
+  GROUP_VARS2 <- mtcars %>% dplyr::select(am, vs) %>% names()
+
+  # Long
+  frequencies <- mtcars %>%
+    dplyr::select(carb, gear, am, vs) %>%
+    cross_freqs(
+      group_vars = GROUP_VARS2, #am & vs
+      exclude_groups = TRUE,
+      include_overall = TRUE
+    )
+  grouping_vars <- unique(frequencies$group_var_name)
+  group_sums <- frequencies %>%
+    dplyr::group_by(group_var_name) %>%
+    dplyr::mutate(sums = sum(n)) %>%
+    dplyr::distinct(group_var_name, .keep_all = TRUE)
+  expect_equal(grouping_vars, c("Overall", "am", "vs"))
+  expect_equal(length(grouping_vars), 3)
+  expect_equal(
+    group_sums$sums[1],
+    group_sums$sums[2],
+    group_sums$sums[3]
+  )
+
+  # Wide
+  frequencies <- mtcars %>%
+    dplyr::select(carb, gear, am, vs) %>%
+    cross_freqs(
+      group_vars = GROUP_VARS2, #am & vs
+      wide = TRUE,
+      exclude_groups = TRUE,
+      include_overall = TRUE
+    )
+  expect_equal(nrow(frequencies), 3)
+  expect_equal(frequencies[[1]][1], 'Overall')
+})
+